🔨Refactor: hook 더 개별화 시키기
