+++ src/components/PokemonSearch.jsx
@@ -1,4 +1,5 @@
 import useDebounce from "../hooks/useDebounce";
+import useToggle from "../hooks/useToggle";
 import {
   PokemonSearchTitle,
   SearchDiv,
@@ -8,8 +9,9 @@ import {
 
 const PokemonSearch = () => {
   // useDebounce 훅을 사용하여 검색어 및 토글 상태를 관리하는 변수와 함수들 가져오기
-  const { inputValue, handleSearchChange, togglePokemonList, showAllPokemon } =
-    useDebounce();
+  const { searchTerm, showAllPokemon, handleSearchChange } = useDebounce();
+  // 토글
+  const togglePokemon = useToggle();
 
   return (
     <SearchDiv>
@@ -20,11 +22,11 @@ const PokemonSearch = () => {
         type="text"
         placeholder="포켓몬을 검색하세요..."
         autoFocus
-        value={inputValue} // 검색어를 입력 값으로 설정
+        value={searchTerm} // 검색어를 입력 값으로 설정
         onChange={handleSearchChange} // 입력 값 변경 시 검색어 업데이트
       />
       {/* 포켓몬 목록 보기/숨기기 토글 버튼 */}
-      <StyledButton className="myPokemonBtn" onClick={togglePokemonList}>
+      <StyledButton className="myPokemonBtn" onClick={togglePokemon}>
         {/* 현재 상태에 따라 버튼 텍스트 변경 */}
         {showAllPokemon ? "나의 포켓몬 보기" : "모든 포켓몬 보기"}
       </StyledButton>
