+++ src/pages/Dex.jsx
@@ -3,36 +3,14 @@ import Dashboard from "../components/Dashboard";
 import PokemonList from "../components/PokemonList";
 import PokemonSearch from "../components/PokemonSearch";
 import { RootDiv, ScrollToTopButton } from "../styles/StyledComponents";
-import { useEffect } from "react";
-import useTopBtn from "../hooks/useTopBtn";
+import useTopBtn from "../util/ScrollTopBtn";
+import useScrollGet from "../hooks/useScrollGet";
 
 const Dex = () => {
   // Redux 상태에서 showAllPokemon 값을 가져옴
   const { showAllPokemon } = useSelector((state) => state.toggle);
   const { useToTop } = useTopBtn();
-
-  // 페이지가 변경될 때 스크롤 위치 저장
-  useEffect(() => {
-    const handleBeforeUnload = () => {
-      sessionStorage.setItem("scrollPosition", 0);
-    };
-
-    window.addEventListener("beforeunload", handleBeforeUnload);
-
-    return () => {
-      window.removeEventListener("beforeunload", handleBeforeUnload);
-    };
-  }, []);
-
-  // 페이지 로드 후 스크롤 위치 복원
-  useEffect(() => {
-    const scrollPosition = sessionStorage.getItem("scrollPosition");
-    if (scrollPosition) {
-      setTimeout(() => {
-        window.scrollTo(0, parseInt(scrollPosition, 10)); // 이전 위치로 스크롤 복원
-      }, 0);
-    }
-  }, []); // 의존성 배열이 빈 배열이므로 컴포넌트가 처음 렌더링될 때만 실행
+  useScrollGet();
 
   return (
     <RootDiv>
